fix(voiceChatUtilities): make it work again and modernize
